test: Add more values to testUtils.ts
Update the testUtils.ts file to include more values for testing various
functions.

Add NaN, Infinity, and non empty object to the invalidInputValues array.

Add weekdayTestValues, isNumberTestValues, and isEmptyObjectTestValues
variables to store arrays of invalid and valid values for testing
isValidWeekday, isNumber, and isEmptyObject functions respectively.

The purpose of this file and variables is to centralize and organize
all testing values in one place within the codebase.