feat: introduce chunkMonthExtended util function
The `chunkMonthExtended` function allows for efficient splitting
of a given `DateRange` into arrays of `DateTime` objects, each representing
a week's worth of dates. This utility is particularly valuable for calendar
implementations, where breaking down a DateRange into weekly chunks is
often necessary for displaying data.

The commit includes the implementation of `chunkMonthExtended` and
a test suite.

chore: add node types to tsconfig